requirements.txt
ajaxuploader==0.3.8
awsebcli==3.9.0
bebop==0.0.6
bleach==1.5.0
blessed==1.9.5
botocore==1.5.1
cement==2.8.2
colorama==0.3.7
defusedxml==0.4.1
Django==1.10.3
django-ajax-selects==1.5.2
django-braces==1.10.0
django-crispy-forms==1.6.1
django-embed-video==1.1.0
django-filebrowser==3.7.2
django-grappelli==2.8.3
django-haystack==2.5.1
django-memcache-status==1.3
django-taggit==0.21.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
easy-thumbnails==2.3
geoip2==2.4.2
html5lib==0.9999999
jmespath==0.9.0
lxml==3.7.1
maxminddb==1.2.2
oauthlib==2.0.0
pathspec==0.5.0
Pillow==3.4.2
pkg-resources==0.0.0
PyJWT==1.4.2
pysolr==3.6.0
python-dateutil==2.6.0
python-memcached==1.58
python-social-auth==0.2.21
python3-openid==3.0.10
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.7.0
semantic-version==2.5.0
six==1.10.0
sorl-thumbnail==12.3
tabulate==0.7.5
twitter-text-python==1.1.0
Unipath==1.1
wcwidth==0.1.7
websocket-client==0.40.0
